Moved the remaining files to Pages. Added subfolder PageUtility.
